Cleaned up some of the interfaces.
Moved PortControlBase definition to router.h and renamed it to PortInterface.
Moved OutputArbitration to be in PortInterface.
