+++ src/sst/elements/merlin/router.h
@@ -27,7 +27,7 @@
 #include <sst/core/unitAlgebra.h>
 #include <sst/core/interfaces/simpleNetwork.h>
 
-using namespace SST;
+#include <queue>
 
 namespace SST {
 namespace Merlin {
@@ -386,7 +386,84 @@ protected:
     Output &output;
 };
 
-class PortControlBase;
+// Class to manage link between NIC and router.  A single NIC can have
+// more than one link_control (and thus link to router).
+class PortInterface : public SubComponent{
+
+public:
+
+    // params are: parent router, router id, port number, topology object
+    SST_ELI_REGISTER_SUBCOMPONENT_API(SST::Merlin::PortInterface, Router*, int, int, Topology*)
+
+    typedef std::queue<internal_router_event*> port_queue_t;
+    typedef std::queue<TopologyEvent*> topo_queue_t;
+
+    virtual void sendTopologyEvent(TopologyEvent* ev) = 0;
+    // Returns true if there is space in the output buffer and false
+    // otherwise.
+    virtual void send(internal_router_event* ev, int vc) = 0;
+    // Returns true if there is space in the output buffer and false
+    // otherwise.
+    virtual bool spaceToSend(int vc, int flits) = 0;
+    // Returns NULL if no event in input_buf[vc]. Otherwise, returns
+    // the next event.
+    virtual internal_router_event* recv(int vc) = 0;
+    virtual internal_router_event** getVCHeads() = 0;
+    
+    // time_base is a frequency which represents the bandwidth of the link in flits/second.
+    PortInterface(ComponentId_t cid) :
+        SubComponent(cid)
+        {}
+
+    PortInterface(Component* parent) :
+        SubComponent(parent)
+        {}
+
+    virtual void initVCs(int vcs, internal_router_event** vc_heads, int* xbar_in_credits, int* output_queue_lengths) = 0;
+
+
+    virtual ~PortInterface() {}
+    // void setup();
+    // void finish();
+    // void init(unsigned int phase);
+    // void complete(unsigned int phase);
+    
+
+    virtual void sendInitData(Event *ev) = 0;
+    virtual Event* recvInitData() = 0;
+    virtual void sendUntimedData(Event *ev) = 0;
+    virtual Event* recvUntimedData() = 0;
+    
+    virtual void dumpState(std::ostream& stream) {}
+    virtual void printStatus(Output& out, int out_port_busy, int in_port_busy) {}
+    
+    // void setupVCs(int vcs, internal_router_event** vc_heads
+	virtual bool decreaseLinkWidth() = 0;
+	virtual bool increaseLinkWidth() = 0; 
+
+
+    class OutputArbitration : public SubComponent {
+    public:
+
+        SST_ELI_REGISTER_SUBCOMPONENT_API(SST::Merlin::PortInterface::OutputArbitration)
+    
+        OutputArbitration(Component* parent) :
+            SubComponent(parent)
+        {}
+        OutputArbitration(ComponentId_t cid) :
+            SubComponent(cid)
+        {}
+        virtual ~OutputArbitration() {}
+
+        virtual void setVCs(int num_vcs) = 0;
+        virtual int arbitrate(PortInterface::port_queue_t* out_q, int* port_out_credits, bool isHostPort, bool& have_packets) = 0;
+        virtual void dumpState(std::ostream& stream) {};
+};
+
+
+
+};
+
 
 class XbarArbitration : public SubComponent {
 public:
@@ -402,9 +479,9 @@ public:
     virtual ~XbarArbitration() {}
 
 #if VERIFY_DECLOCKING
-    virtual void arbitrate(PortControlBase** ports, int* port_busy, int* out_port_busy, int* progress_vc, bool clocking) = 0;
+    virtual void arbitrate(PortInterface** ports, int* port_busy, int* out_port_busy, int* progress_vc, bool clocking) = 0;
 #else
-    virtual void arbitrate(PortControlBase** ports, int* port_busy, int* out_port_busy, int* progress_vc) = 0;
+    virtual void arbitrate(PortInterface** ports, int* port_busy, int* out_port_busy, int* progress_vc) = 0;
 #endif
     virtual void setPorts(int num_ports, int num_vcs) = 0;
     virtual bool isOkayToPauseClock() { return true; }
